+++ src/telliot/utils/rpc_endpoint.py
@@ -1,13 +1,8 @@
 """
-Utils for:
-- creating a JSON RPC connection to an EVM blockchain
-- connecting to an EVM contract
+Utils for creating a JSON RPC connection to an EVM blockchain
 """
 from typing import Optional
-from typing import Union
 
-import web3
-from pydantic.dataclasses import dataclass
 from pydantic.main import BaseModel
 from web3 import Web3
 
@@ -47,26 +42,3 @@ class RPCEndpoint(BaseModel):
             print("Could not connect to {}".format(self))
 
         return self.web3
-
-
-@dataclass
-class Contract:
-    """Convenience wrapper for connecting to an Ethereum contract"""
-
-    #: RPCNode connection to Ethereum network
-    node: RPCEndpoint
-
-    address: str
-
-    #: Abi specifications of contract, likely loaded from JSON
-    abi: str
-
-    def connect(self) -> Union[web3.contract.Contract, TypeError]:
-        """Connect to EVM contract through an RPC Endpoint"""
-        if self.node.web3 is None:
-            return TypeError("node is not connected")
-        else:
-            self.Contract = self.node.web3.eth.contract(
-                address=self.address, abi=self.abi
-            )
-            return self.Contract
