+++ tests/test_rpc_endpoint.py
@@ -1,11 +1,9 @@
 """
-Tests covering Pytelliot ethereum connection utils.
+Tests covering Pytelliot rpc connection  utils.
 """
 import pytest
 import requests
-import web3
 from dotenv import load_dotenv
-from telliot.utils.rpc_endpoint import Contract
 from telliot.utils.rpc_endpoint import RPCEndpoint
 
 load_dotenv()  # we will replace this with loading from config
@@ -44,27 +42,6 @@ def test_incomplete_rpc_url():
     with pytest.raises(requests.exceptions.HTTPError):
         endpt.web3.eth.block_number
 
-
-def test_read_tellor_playground():
-    """Contract object should access Tellor Playground functions"""
-    contract = connect_to_contract("0x4699845F22CA2705449CFD532060e04abE3F1F31")
-    assert len(contract.Contract.all_functions()) > 0
-    assert isinstance(
-        contract.Contract.all_functions()[0], web3.contract.ContractFunction
-    )
-
-
 def test_load_from_config():
     """RPCEndpoint should read from config.yml"""
     pass
-
-
-def connect_to_contract(address):
-    """Helper function for connecting to a contract at an address"""
-    url = "https://mainnet.infura.io/v3/1a09c4705f114af2997548dd901d655b"
-    endpt = RPCEndpoint(network=network, provider=provider, url=url)
-    endpt.connect()
-
-    c = Contract(endpt, address, abi)
-    c.connect()
-    return c
